requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pyserial>=3.5
pydantic>=2.7